errors fixed

+++ coupon-service/src/main/java/smr/shop/coupon/service/dto/request/CouponCreateRequest.java
@@ -5,6 +5,7 @@ import lombok.Value;
 import smr.shop.coupon.service.model.valueobject.CouponDiscountType;
 import smr.shop.coupon.service.model.valueobject.CouponType;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 /**
@@ -29,13 +30,13 @@ public class CouponCreateRequest {
     CouponDiscountType discountType;
 
     @NotBlank
-    Double amount;
+    BigDecimal amount;
 
     @NotBlank
-    Float percentage;
+    short percentage;
 
     @NotBlank
-    Double maxDiscountPrice;
+    BigDecimal maxDiscountPrice;
 
     @NotBlank
     ZonedDateTime endDate;
